+++ BTCPayServer/Services/WalletFileParsing/OutputDescriptorJsonWalletFileParser.cs
@@ -1,36 +1,34 @@
 #nullable enable
 using System;
+using System.Diagnostics.CodeAnalysis;
 using BTCPayServer;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 namespace BTCPayServer.Services.WalletFileParsing;
 public class OutputDescriptorJsonWalletFileParser : IWalletFileParser
 {
     private readonly OutputDescriptorWalletFileParser _outputDescriptorOnChainWalletParser;
 
+    class OutputDescriptorJsonWalletFileFormat
+    {
+        public string? Descriptor { get; set; }
+        public string? Source { get; set; }
+    }
     public OutputDescriptorJsonWalletFileParser(OutputDescriptorWalletFileParser outputDescriptorOnChainWalletParser)
     {
         _outputDescriptorOnChainWalletParser = outputDescriptorOnChainWalletParser;
     }
-    public (DerivationSchemeSettings? DerivationSchemeSettings, string? Error) TryParse(BTCPayNetwork network,
-        string data)
+    public bool TryParse(BTCPayNetwork network, string data, [MaybeNullWhen(false)] out DerivationSchemeSettings derivationSchemeSettings)
     {
-        try
-        {
-            var jobj = JObject.Parse(data);
-            if (!jobj.TryGetValue("Descriptor", StringComparison.InvariantCultureIgnoreCase, out var descriptorToken) ||
-                descriptorToken?.Value<string>() is not string desc)
-                return (null, null);
+        derivationSchemeSettings = null;
+        var jobj = JsonConvert.DeserializeObject<OutputDescriptorJsonWalletFileFormat>(data);
+        if (jobj?.Descriptor is null)
+            return false;
 
-                
-            var result =  _outputDescriptorOnChainWalletParser.TryParse(network, desc);
-            if (result.DerivationSchemeSettings is not null && jobj.TryGetValue("Source", StringComparison.InvariantCultureIgnoreCase, out var sourceToken))
-                result.DerivationSchemeSettings.Source = sourceToken.Value<string>();
-            return result;
-                
-        }
-        catch (Exception)
-        {
-            return (null, null);
-        }
+        if (!_outputDescriptorOnChainWalletParser.TryParse(network, jobj.Descriptor, out derivationSchemeSettings))
+            return false;
+        if (jobj.Source is not null)
+            derivationSchemeSettings.Source = jobj.Source;
+        return true;
     }
 }