+++ Changelog.md
@@ -13,6 +13,10 @@
 * Fix: Checkout v1 was not applying the custom style (#5628 #5615 #5616) @dennisreimann
 * Fix: Test email with multiple recipients was crashing (#5649 #5648) @dennisreimann
 * Fix: Test webhook for payment requests (#5680) @Kukks
+* Fix: Sometimes importing a wallet file from Electrum would fail @NicolasDorier
+* Fix: Creating a Store as a Guest generates a 403 error (#5688 #5689) @dennisreimann
+* Fix: In Wallet Send, label were not applied to transactions (#5700) @NicolasDorier
+* Fix: "View recent invoices" in Keypad PoS should be accessible for authenticated Guest users (#5702 #5698) @dennisreimann
 
 ### Improvements